+++ packages/db/migrations/0012_mushy_alex_power.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `message` ADD `opened_at` integer;--> statement-breakpoint
+UPDATE `message` SET `opened_at` = `created_at` WHERE `opened_at` IS NULL;